* test: introduce the http expect testing tool - split common methods in the test handler - correct the error in the log directory during testing
